Rename common.h to ember_common.h
