+++ ember-bltools/ember_bltools.h
@@ -1,7 +1,7 @@
 #ifndef EMBER_BLTOOLS_H
 #define EMBER_BLTOOLS_H
 
-#include "common.h"
+#include "ember_common.h"
 
 bool ember_bltools_set_boot_partition_to_factory();
