+++ ember-commonh/ember_common.h
