+++ ember-commonh/library.json
@@ -1,5 +1,5 @@
 {
   "name": "ember-commonh",
   "version": "1.0.0",
-  "description": "Ember common.h utility header"
+  "description": "Ember common utility header"
 }
